+++ src/clj/clj_toolbox/derefs.clj
@@ -1,4 +1,5 @@
-(ns clj-toolbox.derefs)
+(ns clj-toolbox.derefs
+  "Functions relatable to clojure.lang.IDeref")
 
 (defn derefable?
   "Returns true if x is an instance of clojure.langIDeref"